ledOps/requirement.txt
paho-mqtt==1.5.1
RPi.GPIO==0.7.0
